+++ packages/hippo_utils/lib/hippo_utils.dart
@@ -37,13 +37,11 @@ export 'src/implementations/content_resolver/get_content_from_uri.dart';
 export 'src/extensions/stream_extensions.dart';
 
 // json_schema
-export 'src/json_schema/json_schema.dart';
 export 'src/json_schema/json_schema_nodes.dart';
 export 'src/json_schema/json_schema_validator.dart';
 export 'src/json_schema/json_schema_path.dart';
 
 // json_pointer
-export 'src/json_pointer/json_pointer.dart';
 export 'src/json_pointer/json_pointer_map.dart';
 
 // json_annotator