+++ packages/hippo_utils/lib/src/json_schema/json_schema_path.dart
@@ -9,8 +9,7 @@
 */
 
 import 'package:flutter/foundation.dart';
-
-import '../json_pointer/json_pointer.dart';
+import 'package:json_schema/json_schema.dart';
 
 @immutable
 sealed class JsonSchemaPathSegment {